Moving istio methods to utilitary class
