+++ butler/src/app/v2/operator/use-cases/reconcile-routes.usecase.ts
@@ -128,32 +128,12 @@ export class ReconcileRoutesUsecase {
       kind: spec.kind,
       status: false
     }
-    if (this.checkEmptySpecs(observed)) {
+    if (ReconcileUtils.checkObservedRoutesEmptiness(observed)) {
       baseResponse.status = false
       return baseResponse
     }
-    baseResponse.status = this.checkComponentExistsOnObserved(observed, spec, circleId)
+    baseResponse.status = ReconcileUtils.checkIfComponentRoutesExistOnObserved(observed, spec, circleId)
     return baseResponse
-  }
-
-  private checkEmptySpecs(observed: PartialRouteHookParams): boolean {
-    const emptyDestinationRules = isEmpty(observed.children['DestinationRule.networking.istio.io/v1alpha3'])
-    const emptyVirtualServices = isEmpty(observed.children['VirtualService.networking.istio.io/v1alpha3'])
-    return emptyDestinationRules || emptyVirtualServices
-  }
-
-  private checkComponentExistsOnObserved(observed: PartialRouteHookParams, spec: SpecsUnion, circleId: string): boolean {
-    const observedDestinationRules = observed.children['DestinationRule.networking.istio.io/v1alpha3'][spec.metadata.name]
-    const observedVirtualService = observed.children['VirtualService.networking.istio.io/v1alpha3'][spec.metadata.name]
-
-    if (!observedDestinationRules || !observedVirtualService) {
-      return false
-    }
-
-    const destinationRulesCircles : string[] = JSON.parse(observedDestinationRules.metadata.annotations.circles)
-    const virtualServiceCircles : string [] = JSON.parse(observedVirtualService.metadata.annotations.circles)
-
-    return destinationRulesCircles.includes(circleId) && virtualServiceCircles.includes(circleId)
   }
 
   private getServicesWithMetadata(components: ComponentEntityV2[]): KubernetesManifest[] {
